+++ cfg/conf.sample.php
@@ -22,10 +22,6 @@ fileupload = false
 ; preselect the burn-after-reading feature, defaults to false
 burnafterreadingselected = false
 
-; delete a burn after reading paste immediatly after it is first accessed from
-; the server and do not wait for a successful decryption
-instantburnafterreading = false
-
 ; which display mode to preselect by default, defaults to "plaintext"
 ; make sure the value exists in [formatter_options]
 defaultformatter = "plaintext"
@@ -84,10 +80,6 @@ zerobincompatibility = false
 ; expire value that is selected per default
 ; make sure the value exists in [expire_options]
 default = "1week"
-
-; optionally the "clone" button can be disabled on expiring pastes
-; note that this only hides the button, copy & paste is still possible
-; clone = false
 
 [expire_options]
 ; Set each one of these to the number of seconds in the expiration period,